requirements.txt
pywinauto>=0.6.8
pyautogui>=0.9.54
pillow>=10.0.0
keyboard>=0.13.5
mouse>=0.7.1
pyyaml>=6.0
pandas>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pywin32>=306
openpyxl>=3.1.0
xlrd>=2.0.0

